backend/requirements.txt
openai~=0.28.0
fastapi~=0.103.1
gTTS~=2.3.2
python-multipart~=0.0.6
ffmpeg-python~=0.2.0
pydantic~=2.3.0
uvicorn~=0.23.2
pydub~=0.25.1
requests~=2.31.0
edge-tts~=6.1.8
elevenlabs~=0.2.26
google-cloud-texttospeech~=2.16.3
google-cloud-speech~=2.25.1
groq~=0.4.2
pyht~=0.0.26
deep-translator
requests